+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 ## Unreleased
 
+### Features
+
+- New Sentry KMP Gradle plugin ([#230](https://github.com/getsentry/sentry-kotlin-multiplatform/pull/230))
+  - Install via `plugins { id("io.sentry.kotlin.multiplatform.gradle") version "{version}" }`  
+  - Enables auto installing of the KMP SDK to commonMain (if all targets are supported)
+  - Enables auto installing of the required Sentry Cocoa SDK with Cocoapods (if Cocoapods plugin is enabled)
+  - Configures linking for SPM (needed if you want to compile a dynamic framework)
+  - Configure via the `sentryKmp` configuration block in your build file
+
 ### Dependencies
 
 - Bump Kotlin version from v1.9.21 to v1.9.23 ([#250](https://github.com/getsentry/sentry-kotlin-multiplatform/pull/250))